+++ h2integrate/converters/methanol/smr_methanol_plant.py
@@ -112,7 +112,7 @@ class SMRMethanolPlantCostModel(MethanolCostBaseClass):
         electricity: hourly electricity consumption (kW*h/h)
         meoh_syn_cat_price: price of methanol synthesis catalyst (USD/ft**3)
         meoh_atr_cat_price: price of methanol ATR catalyst (USD/ft**3)
-        ng_price: price of NG (USD/MBtu)
+        ng_price: price of NG (USD/MMBtu)
     Outputs:
         meoh_syn_cat_cost: annual cost of methanol synthesis catalyst (USD/year)
         meoh_atr_cat_cost: annual cost of methanol ATR catalyst (USD/year)
@@ -133,9 +133,7 @@ class SMRMethanolPlantCostModel(MethanolCostBaseClass):
         self.add_input("electricity", shape=8760, units="kW*h/h")
         self.add_input("meoh_syn_cat_price", units="USD/ft**3", val=self.config.meoh_syn_cat_price)
         self.add_input("meoh_atr_cat_price", units="USD/ft**3", val=self.config.meoh_atr_cat_price)
-        self.add_input(
-            "ng_price", units="USD/MBtu", val=self.config.ng_price
-        )  # TODO: get OpenMDAO to recognize 'MMBtu'
+        self.add_input("ng_price", units="USD/MMBtu", val=self.config.ng_price)
 
         self.add_output("meoh_syn_cat_cost", units="USD/year")
         self.add_output("meoh_atr_cat_cost", units="USD/year")
@@ -150,7 +148,7 @@ class SMRMethanolPlantCostModel(MethanolCostBaseClass):
         ppa_price = self.options["plant_config"]["plant"]["ppa_price"]
 
         lhv_mj = inputs["ng_lhv"]
-        lhv_mmbtu = convert_units(lhv_mj, "MJ", "MBtu")
+        lhv_mmbtu = convert_units(lhv_mj, "MJ", "MMBtu")
 
         outputs["CapEx"] = toc_usd
         outputs["OpEx"] = foc_usd_y + voc_usd_y